Restructured files for refactoring using Lazy Loading class provided by Click, added back (tool).py files within each tool subdirectory to allow for commands to be housed under respective tool groups, while getting rid of almost all the code in fre.py. Now commands and groups should only be loaded when needed
