+++ fre/cmor/__init__.py
@@ -1,3 +1,3 @@
-from .CMORmixer import run_subtool
+from .CMORmixer import cmor_run_subtool
 
-__all__ = ["run_subtool"]
+__all__ = ["cmor_run_subtool"]
