--- fre/fake/gfdlfremake/schema.json
@@ -1,164 +1,0 @@
-{
-	"$schema": "https://json-schema.org/draft/2020-12/schema",
-  	"$id": "https://example.com/product.schema.json",
-	"title": "frecanopy", 
-	"description": "A yaml base for use with fre canopy",
-	"type": "object",
-	"properties": {
-		"platformYaml": {
-                        "description": "Path to the platform yaml file",
-                        "type": "string"
-		},
-		"compileYaml": {
-                        "description": "Path to the compile yaml file",
-                        "type": "string"
-                },
-		"experiment": {
-			"description": "The name of the model",
-			"type": "string"
-		},
-                "container_addlibs":{
-                        "description": "Libraries and packages needed for linking in the container",
-                        "type": ["array","string","null"]
-                },
-                "baremetal_linkerflags":{
-                        "description": "Linker flags of libraries and packages needed for linking in the bare-metal build",
-                        "type": ["array","string","null"]
-                },
-		"src":{
-			"description": "The source code descriptions",
-			"type": "array",
-			"items":{"$ref": "#/$defs/comp" }
-		},
-		"platforms":{
-                        "description": "FRE platforms",
-                        "type": "array",
-                        "items":{"$ref": "#/$defs/plat" }
-                }
-	},
-	"$defs":{
-		"comp": {
-			"type": "object",
-			"required": ["component", "repo"],
-			"properties":{
-				"component": {
-					"type": "string",
-					"description": "The name of the model component"
-				},
-				"repo": {
-					"anyOf": [
-						{ 
-							"type": "string",
-                                        		"description": "The URL of the code repository"
-						},
-						{
-                                                        "type": ["array","string"],
-                                                        "description": "The URL of the code repository"
-                                                }
-						]
-				},
-				"branch": {
-					"anyOf": [
-                                                {
-                                                        "type": "string",
-                                                        "description": "The version of code to clone"
-                                                },
-                                                {
-                                                        "type": ["array","string"],
-                                                        "description": "The version of code to clone"
-                                                }
-                                                ]
-				},
-				"requires": {
-                                        "type": ["array","string"],
-                                        "description": "list of componets that this component depends on"
-				},
-				"cppdefs": {
-                                        "type": "string",
-                                        "description": "String of CPPDEFs to include in compiling the component"
-                                },
-				"paths": {
-                                        "type": ["array","string"],
-                                        "description": "A list of the paths in the component to compile"
-                                },
-				"additionalInstructions": {
-                                        "type": "string",
-                                        "description": "Additional instuctions to run after checkout"
-				},
-				"doF90Cpp": {
-                                        "type": "boolean",
-                                        "description": "True if the preprocessor needs to be run"
-                                },
-                                "makeOverrides": {
-                                        "type": "string",
-                                        "description": "Overrides openmp target for MOM6" 
-                                },
-                                "otherFlags": {
-                                        "type": "string",
-                                        "description": "String of Include flags necessary to retrieve other code needed"
-                                }
-			},
-			"additionalProperties": false
-		},
-		"plat": {
-                        "type": "object",
-                        "required": ["name", "compiler", "fc", "cc"],
-                        "properties":{
-				"name": {
-                                        "type": "string",
-                                        "description": "The name of the platform"
-				},
-				"compiler": {
-                                        "type": "string",
-                                        "description": "The compiler used to build the model"
-				},
-				"modulesInit": {
-                                        "type": ["array","string"],
-                                        "description": "Array of commands to run before loading modules"
-				},
-				"modules": {
-                                        "type": ["array","string"],
-                                        "description": "List (array) of modules to load"
-				},
-				"fc": {
-                                        "type": "string",
-                                        "description": "The Fortran compiler"
-				},
-				"cc": {
-                                        "type": "string",
-                                        "description": "The C compiler"
-                                },
-				"mkTemplate": {
-                                        "type": "string",
-                                        "description": "Path to the mk template file"
-				},
-				"modelRoot": {
-                                        "type": "string",
-                                        "description": "Path to the root for all model install files"
-				},
-				"container": {
-					"anyOf": 
-						[
-							{"type": "string"},
-							{"type": "boolean"}
-						]
-				},
-				"RUNenv": {
-					"type": ["array", "string"],
-					"description": "Commands needed at the beginning of a RUN in dockerfile"
-				},
-				"containerBuild": {
-                                        "type": "string",
-                                        "description": "Program used to build the container"
-				},
-				"containerRun": {
-                                        "type": "string",
-                                        "description": "Program used to run the container"
-				}
-			},
-                        "additionalProperties": false
-                }
-	},
-	"required": ["experiment", "src", "platforms"],
-	"additionalProperties": true
-}
