--- fre/fake/gfdlfremake/yamls/compile.yaml
@@ -1,66 +1,0 @@
-experiment: "am5"
-container_addlibs: 
-baremetal_linkerflags:  
-src:
-     - component: "FMS"
-       repo: "https://github.com/NOAA-GFDL/FMS.git"
-       cppdefs: "-DINTERNAL_FILE_NML -Duse_libMPI -Duse_netCDF" 
-       branch: "2022.01"
-       cppdefs: "-DHAVE_GETTID -Duse_libMPI -Duse_netCDF"
-       otherFlags: "$(FMSincludes)"
-     - component: "am5_phys"
-       requires: ["FMS"]
-       repo: "https://gitlab.gfdl.noaa.gov/FMS/am5_phys.git"
-       branch: "2022.01"
-       otherFlags: "$(FMSincludes)"
-     - component: "GFDL_atmos_cubed_sphere"
-       requires: ["FMS", "am5_phys"]
-       repo: "https://github.com/NOAA-GFDL/GFDL_atmos_cubed_sphere.git"
-       cppdefs: "-DSPMD -DCLIMATE_NUDGE -DINTERNAL_FILE_NML"
-       branch: "2022.01"
-       paths: ["GFDL_atmos_cubed_sphere/driver/GFDL",
-              "GFDL_atmos_cubed_sphere/model",
-              "GFDL_atmos_cubed_sphere/driver/SHiELD/cloud_diagnosis.F90",
-              "GFDL_atmos_cubed_sphere/driver/SHiELD/gfdl_cloud_microphys.F90",
-              "GFDL_atmos_cubed_sphere/tools",
-              "GFDL_atmos_cubed_sphere/GFDL_tools"]
-       otherFlags: "$(FMSincludes)"
-     - component: "atmos_drivers"
-       requires: ["FMS", "am5_phys", "GFDL_atmos_cubed_sphere"]
-       repo: "https://github.com/NOAA-GFDL/atmos_drivers.git"
-       cppdefs: "-DSPMD -DCLIMATE_NUDGE"
-       branch: "2022.01"
-       paths: ["atmos_drivers/coupled"]
-       otherFlags: "$(FMSincludes)"
-     - component: "ice_sis"
-       requires: ["FMS", "ice_param", "mom6"]
-       repo: "https://gitlab.gfdl.noaa.gov/FMS/ice_sis.git"
-       branch: "2021.02"
-       otherFlags: "$(FMSincludes) $(momIncludes)"
-     - component: "ice_param"
-       repo: "https://github.com/NOAA-GFDL/ice_param.git"
-       cppdefs: "-Duse_yaml -Duse_libMPI -Duse_netCDF"
-       branch: "2021.02"
-       requires: ["FMS", "mom6"]
-       otherFlags: "$(FMSincludes) $(momIncludes)"
-     - component: "land_lad2"
-       requires: ["FMS"]
-       repo: "https://gitlab.gfdl.noaa.gov/FMS/land_lad2.git"
-       branch: "2022.01"
-       branch: "land_lad2_2021.02"
-       doF90Cpp: True
-       cppdefs: "-DINTERNAL_FILE_NML"
-       otherFlags: "$(FMSincludes)"
-     - component: "mom6"
-       requires: ["FMS"]
-       paths: ["mom6/MOM6-examples/src/MOM6/config_src/dynamic", "mom6/MOM6-examples/src/MOM6/config_src/coupled_driver", "mom6/MOM6-examples/src/MOM6/src/*/", "mom6/MOM6-examples/src/MOM6/src/*/*/", "mom6/ocean_BGC/generic_tracers", "mom6/ocean_BGC/mocsy/src"]
-       branch: ["2021.02","dev/gfdl/2018.04.06"]
-       repo: ["https://github.com/NOAA-GFDL/ocean_BGC.git","https://github.com/NOAA-GFDL/MOM6-examples.git"]
-       makeOverrides: 'OPENMP=""'
-       otherFlags: "$(FMSincludes) $(momIncludes)"
-     - component: "FMScoupler"
-       paths: ["FMScoupler/full", "FMScoupler/shared"]
-       repo: "https://github.com/NOAA-GFDL/FMScoupler.git"
-       branch: "2022.01"
-       requires: ["FMS", "atmos_drivers", "am5_phys", "land_lad2", "ice_sis", "ice_param", "mom6"]
-       otherFlags: "$(FMSincludes) $(momIncludes)"
