--- fre/fake/gfdlfremake/yamls/schema.json
@@ -1,1 +1,0 @@
-../schema.json
