+++ fre/lazy_group.py
@@ -25,7 +25,7 @@ class LazyGroup(click.Group):
         import_path = self.lazy_subcommands[cmd_name]
         modname, cmd_object_name = import_path.rsplit(".", 1)
         # do the import
-        mod = importlib.import_module(modname)
+        mod = importlib.import_module(modname, package="fre")
         # get the Command object from that module
         cmd_object = getattr(mod, cmd_object_name)
         # check the result to make debugging easier
