+++ fre/list/frelist.py
@@ -1,0 +1,16 @@
+import click
+from .frelistexample import list_test_function
+
+@click.group(help="- access fre list subcommands")
+def listCli():
+    pass
+
+@listCli.command()
+@click.option('--uppercase', '-u', is_flag=True, help = 'Print statement in uppercase.')
+@click.pass_context
+def function(context, uppercase):
+    """ - Execute fre list test """
+    context.forward(list_test_function)
+
+if __name__ == "__main__":
+    listCli()
