+++ fre/pp/__init__.py
@@ -3,7 +3,7 @@ from .configureScriptYAML import yamlInfo
 from .configureScriptXML import convert
 from .validate import validate_subtool
 from .install import install_subtool
-from .run import run_subtool
+from .run import pp_run_subtool
 from .status import status_subtool
 
 __all__ = ["checkoutTemplate",
@@ -11,6 +11,6 @@ __all__ = ["checkoutTemplate",
            "convert",
            "validate_subtool",
            "install_subtool",
-           "run_subtool",
+           "pp_run_subtool",
            "status_subtool"]
